chore: Refactor App component and Card component
Refactor the App component to use the useState hook to manage the state of the 'age' variable. Add a 'hello' variable and a 'user' object to the state. Update the Card component to destructure the 'user' prop. Map the 'usersData' array to render multiple Card components. Add functions to increase and decrease the age, as well as clear the 'usersData' array.

Fixes #123